+++ elasticapm/conf/constants.py
@@ -1,4 +1,4 @@
-EVENTS_API_PATH = "/intake/v2/events"
+EVENTS_API_PATH = "intake/v2/events"
 
 TRACE_CONTEXT_VERSION = 0
 TRACEPARENT_HEADER_NAME = "elastic-apm-traceparent"