+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # Changelog
 
 ## Unreleased
+[Check the diff](https://github.com/elastic/apm-agent-python/compare/v4.0.3...master)
+
+ * Added support for collecting system and process metrics (#361)
+ * Added `transaction.sampled` to errors (#371)
+ * Added parsing of `/proc/self/cgroup` to capture container meta data (#352)
+
+## v4.0.3
+[Check the diff](https://github.com/elastic/apm-agent-python/compare/v4.0.2...v4.0.3)
 
  * implemented de-dotting of tag names and context keys (#353)
+ * wrote a quickfix for the boto3/botocore instrumentation (#367)
+ * fixed an issue with psycopg2 and encoded strings (#366)
 
 ## v4.0.2
 [Check the diff](https://github.com/elastic/apm-agent-python/compare/v4.0.1...v4.0.2)